Fixed being able to overwrite the lower nybble of the Joypad register.
This fixes SGB detection for games that have it. With this fix, I can
now get to the title screen of every game I have, save only the Oracle
games and Zerd no Densetsu! In addition, this also got rid of the
flickering in Kirby's Dream Land, Metroid II, and Ekkusu. I can't
believe that it was this which was breaking almost every game.