Emulate DMG STAT interrupt bug.
On DMG, writing to the STAT register while the LCD is on and in
either mode 1 or 0, bit 1 of the IF register is set and a STAT interrupt
is requested. This bug does not occur on CGB.

After this fix, both Road Rash and Zerd no Densetsu get in game.