CGB - Run emulation at 69920 cycles per frame instead of 70224.
The Game Boy executes 70224 cycles per frame at 59.7275 FPS. To run at
the correct speed, we should be executing 69905 cycles per frame, but
this gives us a sample rate which is very difficult to resample to 800
samples per frame.

Running emulation at 69920 cycles per frame is very close to the correct
speed, and gives us a sample rate of 34960 which is much nicer to
resample.

With this change,
* emulation now runs at very close to the correct speed.
* we no longer have to pre-downsample audio before resampling it.
* our resample buffer is 30% smaller.