GBA - Sample audio at a fixed rate.
This is basically a rewrite of almost all of the audio code. The new
code now samples audio at a fixed rate of 34960 samples per frame,
to match the sample rate of the CGB channels. When we generate a sample,
we play whatever was last popped from the FIFO.

This refactor includes a number of new improvements as well:
* Silence is now generated whenever audio is disabled, and a new output
buffer is generated every frame even when audio is disabled. Now samples
that were already generated in the frame when audio was disabled will
still be played, and audio will resume at the correct time when audio is
re-enabled. This also prevents the previous output buffer from being
replayed while audio is disabled.
* Audio is now submitted to SDL as soon as it is resampled, instead of
right before we wait for VBLANK.
* Using two timers with different sample rates for each FIFO is now
supported (but I've yet to see a game actually do this).
* GBA emulation now runs at very close to the correct speed of 59.7275
FPS instead of 60 FPS, matching the behaviour introduced in commit
995a7b1.