Do biquad filtering on floats instead of doubles.
There is a slight difference in the output samples with this change,
but on GBA it's usually only ~37% of the samples are different, and in
most cases it's within +/-1 of the result computed with doubles.
The largest difference I saw was +/-4, but that was infrequent.

In any case, I can't hear any difference.