Filter two samples at once by executing both biquad passes in parallel.
With this change, the coefficients for two biquads are packed into a
single Biquad object by storing the first biquad in the low half of the
Vec4fs, and the second biquad in the high half. We can then filter two
stereo samples in parallel by filtering one through the first biquad
pass and the other through the second. This did require making the
LowPassFilter function more complex, but the result is a sizeable speed
boost.

This also changed the GBA backend to use a two-pass filter instead of a
three-pass.