GBA - Add PSG channel support.
This required a significant refactoring of the channel code so I'd be
happy with it and so it would mesh better with the rest of the GBA
code.

The Channel class is now templated on the kind of generator it is. I
didn't want to use inheritance, but wasn't happy with the runtime
conditionals it was using before for channel-specific behaviour. Now it
uses template specialization and `if constexpr` for channel-specific
behaviour.

There is also a bunch of cleanup: got rid of magic numbers wherever
possible, added register access functions, etc.